chore: Async user avatar upload processor
